Remove legacy core  API
