--- packages/roosterjs-editor-adapter/lib/coreApi/coreApiMap.ts
@@ -1,9 +1,0 @@
-import { insertNode } from './insertNode';
-import type { EditorAdapterCoreApiMap } from '../publicTypes/EditorAdapterCore';
-
-/**
- * @internal
- */
-export const coreApiMap: EditorAdapterCoreApiMap = {
-    insertNode,
-};
