+++ packages/roosterjs-editor-adapter/lib/index.ts
@@ -1,8 +1,3 @@
-export {
-    EditorAdapterCore,
-    EditorAdapterCoreApiMap,
-    InsertNode,
-} from './publicTypes/EditorAdapterCore';
 export { EditorAdapterOptions } from './publicTypes/EditorAdapterOptions';
 export { BeforePasteAdapterEvent } from './publicTypes/BeforePasteAdapterEvent';
