+++ packages/roosterjs-editor-adapter/lib/publicTypes/EditorAdapterOptions.ts
@@ -1,5 +1,4 @@
 import type { StandaloneEditorOptions } from 'roosterjs-content-model-types';
-import type { EditorAdapterCoreApiMap } from './EditorAdapterCore';
 import type { EditorPlugin, ExperimentalFeatures } from 'roosterjs-editor-types';
 
 /**
@@ -11,12 +10,6 @@ export interface EditorAdapterOptions extends StandaloneEditorOptions {
      * Default value is whatever already inside the editor content DIV
      */
     initialContent?: string;
-
-    /**
-     * A function map to override default core API implementation
-     * Default value is null
-     */
-    legacyCoreApiOverride?: Partial<EditorAdapterCoreApiMap>;
 
     /**
      * Specify the enabled experimental features
